requirements.txt
affine==2.4.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
branca==0.6.0
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.1.3
contextily==1.3.0
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.3
Fiona==1.9.3
Flask==2.3.2
flit_core==3.8.0
folium==0.14.0
fonttools==4.39.3
GDAL==3.6.3
geographiclib==1.52
geopandas==0.12.2
geopy==2.3.0
gtfs-kit==5.2.6
gtfsutils==0.0.5
h3==3.7.6
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json2html==1.3.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter-events==0.6.3
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
mapclassify==2.5.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mercantile==1.2.1
mistune==2.0.5
munch==2.5.0
munkres==1.1.4
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.24.3
openpyxl==3.1.2
osmium==3.6.0
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pip==23.1.1
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
pooch==1.7.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycountry==22.3.5
pycparser==2.21
Pygments==2.15.1
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyproj==3.5.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
rasterio==1.3.6
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Rtree==1.0.1
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.0
setuptools==67.7.1
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
snuggs==1.4.7
soupsieve==2.3.2.post1
stack-data==0.6.2
terminado==0.17.1
threadpoolctl==3.1.0
timezonefinder==6.2.0
tinycss2==1.2.1
tornado==6.3
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzwhere==3.0.3
urllib3==1.26.15
utm==0.7.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.4
wheel==0.40.0
widgetsnbextension==4.0.7
xyzservices==2023.2.0
zipp==3.15.0
